[Decomp][Spell] Cohesion split: extract SpellAuraControl handlers to SpellAuraControl.cpp
